Moved map related files into ui or map package
